Multi-register Tensors.
As part of our design, moves into vector registers now take
a vector_shape. Layouts described layout inside each register.

Part of implementing this involved adding the `address_root`
property, which returns a reference the parent Tensor in a
non-cache memory. This is now used to determine the
contiguousness of tiles.

Minor Changes:
- Removed `height` and `width` getters.
- TileOutAction no longer consumes a callable, just the Impl.
- _mutating_replace now checks type equality, not `isinstance`.