Added verification feature to frontend.
With the new project structure, the verification feature is accidentally
always disabled in frontend/src/main.rs, meaning that we never perform
a correctness check.  This patch adds a verification feature to the
frontend workspace as well and propagates it to the morello workspace's
verification feature so that we can check correctness on the frontend.