Fixed ImplReducer.
Sort Policy:

With this implementation, `ImplReducer` sorts the results by `cost` and
then by `action_idx`.

Time Complexity:

For the `insert` method, both previous and current implementations have
a time complexity of `O(K)`, where K is the value of `top_k`.  This is
because we need to find the same cost for preferences AFAIK, which can
take `O(K)` in the worst case. For the `finalize` method, the previous
implementation has a time complexity of `O(1)`, while the current
implementation takes `O(K)` to convert from a `BTreeSet` to a `SmallVec`
since `BTreeSet`'s iterator is sorted.  Although there is an increase in
time complexity for `finalize`, I believe it will be negligible as we
later require `O(K)` when putting finalized results in DB.

Data Structure:

I chose to use a `BTreeSet` instead of `BinaryHeap` because we cannot
avoid an `O(K)` time complexity with preferences, as explained
previously.  Considering that using `BinaryHeap` could result in an
`O(K log K)` time complexity in the `finalize` method, I believe that
using `BTreeSet` is better.

Space Complexity & Heap Allocation Cost:

It is clear that using `BTreeSet` leads to worse space complexity and
heap allocation costs compared to `SmallVec`. Unlike `SmallVec` and
`BinaryHeap`, `BTreeSet` does not have a `with_capacity` method.
With larger values of K, there might be regression due to increased
heap allocation.

cf. https://github.com/samkaufman/morello/pull/38#discussion_r1443531624