fix an unnecessary Move sub-Spec introduction.
Action::Move would incorrectly introduce a Move sub-Spec when moving into a cache with a layout which would canonicalize to the original layout. This resulted from comparing the original, canonicalized layout to the non-canonical destination layout. This modifies behavior to first canonicalize the destination layout (independent of the contig.), then compare.